Added an update button that gets a weather data for Yandex Weather
I edited a HttpServer that now retrives a dictionary instead of a string, created a button that updates a Yandex Weather Item and now, when you click on a button, hard coded text value changes to an actual weather. Now Conditions model is fully Observable and can be easily changed. New package "Newtonsoft.Json" for unpacking a json sent from a Yandex Weather API.